requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
flake8==7.1.2
gunicorn==23.0.0
iniconfig==2.0.0
isort==6.0.0
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
